requirements.txt
mysql-connector-python>=8.0.0
bcrypt>=4.0.0
PyQt6>=6.4.0
openpyxl>=3.0.0
python-dateutil>=2.8.0
python-barcode>=0.16.1
Pillow>=10.0.0
